refactor: Templates to use Visitor
