+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod enumeration;
+pub mod template;
 pub mod evaluator;
 pub mod lexer;
 pub mod parser;
